updated home image and color pallete
